Plan Snapshots Page
A page can be accessed at /plan-snapshots which contains a brief explanation of what the page is and why there are weekly snapshots, followed by empty links to each snapshot for each week of the project.

All links are currently non-functional, and will be filled out each week as we go along.